requirements.txt
torch==2.1.0
numpy==1.26.0
librosa==0.10.1
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
tqdm==4.66.1
PyYAML==6.0.1